+++ admin/events.php
@@ -67,6 +67,7 @@ function _event_type_de(string $type): string {
 
     'admin_suggestions_imported' => 'Vorschlagskatalog importiert',
     'admin_database_backup_downloaded' => 'Datenbanksicherung heruntergeladen',
+    'teacher_backup_downloaded' => 'Lehrkraft-Datensicherung heruntergeladen',
   ];
   return $map[$type] ?? $type;
 }